[BE] 최근 채팅내역 조회 및 캐싱 구현 / 서비스 서버 이벤트 발행 및 캐싱 구현 #104 (#190)
* feat: dotenv 설정

* feat: application.yml 추가

* refactor: 공통 모듈 사용하도록 변경

* fix: 공통 모듈 config 임포트 하도록 수정

* fix: yml 수정

* fix: 몽고 yml 수정

* fix: 팩토리에 잘못된 dto 타입 주입 수정

* test: 테스트용 로그 생성

* chore: gradle 엘라스틱 버전 변경

* chore: gradle 엘라스틱 버전 변경

* refactor: 엘라스틱 서치 쿼리 형식 변경

* feat: 과거 메시지 조회 기능 구현

* feat: 과거 메시지 조회 기능 구현

* docs: redisKeys 설계 추가

* feat: 서버 채널별 접속정보 저장 로직 구현

* feat: 서버 채널 목록 캐싱 및 채널 생성/수정/삭제 이벤트 처리 구현

* fix: 잘못 된 기존 캐싱 로직 삭제

* refactor: 기존 RedisTTL을 분리

* feat: 서버 목록 캐싱 기능 구현

* feat: 서버 이벤트시 레디스 캐싱 및 카프카에 이벤트 발행하도록 로직 구현, 서버 조회시 없으면 캐싱하는 로직 구현

* fix: 공통 모듈 변경에 따른 수정

* feat: 카테고리 변경시 연관된 채널 정보 변경 이벤트 발행

---------

+++ src/backend/common-module/src/main/java/com/bbebig/commonmodule/kafka/dto/serverEvent/ServerChannelEventDto.java
@@ -1,5 +1,6 @@
 package com.bbebig.commonmodule.kafka.dto.serverEvent;
 
+import com.bbebig.commonmodule.kafka.dto.model.ChannelType;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.experimental.SuperBuilder;
@@ -17,7 +18,7 @@ public class ServerChannelEventDto extends ServerEventDto {
 
 	String channelType; // CHAT, VOICE
 
-	Long order; // Channel 순서, DELETE 시에는 null
+	int order; // Channel 순서, DELETE 시에는 null
 
 	String status; // CREATE, UPDATE, DELETE
